Utilities class to namespace
